Tabulate Adverse Events

The function tabulates adverse events. One AE per ID will be counted in the resulting table. If a by= variable is passed and a patient experienced more than one of the same AE, the AE associated with the highest by= level will be included. For example, if a patient has two of the same AE and by = grade, the AE with the highest grade will be included. Similarly, if tabulations within system organ class are requested, the AE within SOC associated with the highest grade will be tabulated.

Usage

tbl_ae(
  data,
  id,
  ae,
  soc = NULL,
  by = NULL,
  strata = NULL,
  id_df = NULL,
  by_values = NULL,
  missing_text = "Unknown",
  statistic = "{n} ({p})",
  header_by = NULL,
  header_strata = NULL,
  zero_symbol = "—"
)

Arguments

data

Data frame

id

Variable name of the patient ID

ae

Variable name of the adverse event column

soc

Variable name of the system organ class column

by

Variable to split results by, e.g. report AEs by grade

strata

Variable to stratify results by, e.g. report AEs summaries by treatment group

id_df

Optional data frame of complete id values and strata to achieve correct base n for the situation in which not all subjects experience adverse events

by_values

Optional vector of complete by values, listed in desired order, to achieve correct table structure for the situation in which an adverse event of a certain grade is not observed for a given soc

missing_text

String that will be shown for missing levels of by=, Default is "Unknown"

statistic

String indicating the statistics that will be reported. The default is "{n} ({p})"

header_by

String indicating the header to be placed in the table. Default is "**{level}**" and only used when by= is specified.

header_strata

String indicating the strata header to be placed in the table. Default is "**{level}**, N = {n}" and only used when strata= is specified.

zero_symbol

String used to represent cells with zero counts. Default is the em-dash ("\U2014"). Using zero_symbol = NULL will print the zero count statistics, e.g. "0 (0)"

Examples

# Example 1 -----------------------------------------------------------------
df_adverse_events %>%
  tbl_ae(
    id = patient_id,
    ae = adverse_event,
    soc = system_organ_class,
    by = grade,
    strata = trt,
    header_by = "**Grade {level}**"
  ) %>%
  as_kable() # UPDATE THIS WITH PROPER gt image at some point.
#> 
#> 
#> |**Adverse Event**                    |**Grade 1** |**Grade 2** |**Grade 3** |**Grade 4** |**Grade 5** |**Grade 1** |**Grade 2** |**Grade 3** |**Grade 4** |**Grade 5** |
#> |:------------------------------------|:-----------|:-----------|:-----------|:-----------|:-----------|:-----------|:-----------|:-----------|:-----------|:-----------|
#> |Blood and lymphatic system disorders |—           |1 (33)      |—           |1 (33)      |1 (33)      |—           |—           |—           |1 (14)      |6 (86)      |
#> |Anaemia                              |—           |—           |1 (33)      |1 (33)      |—           |—           |—           |1 (14)      |1 (14)      |3 (43)      |
#> |Increased tendency to bruise         |—           |—           |—           |1 (33)      |—           |—           |—           |—           |3 (43)      |2 (29)      |
#> |Iron deficiency anaemia              |—           |—           |—           |1 (33)      |1 (33)      |1 (14)      |2 (29)      |—           |1 (14)      |1 (14)      |
#> |Thrombocytopenia                     |—           |1 (33)      |—           |1 (33)      |—           |—           |—           |3 (43)      |—           |4 (57)      |
#> |Gastrointestinal disorders           |—           |—           |—           |2 (67)      |1 (33)      |—           |—           |—           |2 (29)      |5 (71)      |
#> |Difficult digestion                  |—           |—           |—           |3 (100)     |—           |1 (14)      |—           |—           |—           |1 (14)      |
#> |Intestinal dilatation                |1 (33)      |—           |—           |—           |—           |1 (14)      |1 (14)      |—           |—           |1 (14)      |
#> |Myochosis                            |—           |2 (67)      |1 (33)      |—           |—           |—           |1 (14)      |—           |1 (14)      |3 (43)      |
#> |Non-erosive reflux disease           |3 (100)     |—           |—           |—           |—           |1 (14)      |—           |—           |3 (43)      |3 (43)      |
#> |Pancreatic enzyme abnormality        |—           |—           |1 (33)      |1 (33)      |1 (33)      |2 (29)      |1 (14)      |1 (14)      |1 (14)      |—           |

# Example 2 -----------------------------------------------------------------
df_adverse_events %>%
  tbl_ae(
    id = patient_id,
    ae = adverse_event,
    by = grade,
    header_by = "**Grade {level}**"
  ) %>%
  as_kable() # UPDATE THIS WITH PROPER gt image at some point.
#> 
#> 
#> |**Adverse Event**             |**Grade 1** |**Grade 2** |**Grade 3** |**Grade 4** |**Grade 5** |
#> |:-----------------------------|:-----------|:-----------|:-----------|:-----------|:-----------|
#> |Anaemia                       |—           |—           |2 (20)      |2 (20)      |3 (30)      |
#> |Difficult digestion           |1 (10)      |—           |—           |3 (30)      |1 (10)      |
#> |Increased tendency to bruise  |—           |—           |—           |4 (40)      |2 (20)      |
#> |Intestinal dilatation         |2 (20)      |1 (10)      |—           |—           |1 (10)      |
#> |Iron deficiency anaemia       |1 (10)      |2 (20)      |—           |2 (20)      |2 (20)      |
#> |Myochosis                     |—           |3 (30)      |1 (10)      |1 (10)      |3 (30)      |
#> |Non-erosive reflux disease    |4 (40)      |—           |—           |3 (30)      |3 (30)      |
#> |Pancreatic enzyme abnormality |2 (20)      |1 (10)      |2 (20)      |2 (20)      |1 (10)      |
#> |Thrombocytopenia              |—           |1 (10)      |3 (30)      |1 (10)      |4 (40)      |