Report Values from gtreg tables in-line

Function allows users to report formatted and styled results from gtreg tables in-line.

Usage

# S3 method for tbl_ae
inline_text(x, row, column = NULL, ...)

# S3 method for tbl_ae_count
inline_text(x, row, column = NULL, ...)

# S3 method for tbl_ae_focus
inline_text(x, row, column = NULL, ...)

Arguments

x

an object of class tbl_ae(), tbl_ae_count(), tbl_ae_focus()

row

string indicating the AE or SOC to report

column

column name of cell to report. Use show_header_names(x) to print all column names beside the current header.

...

not used

Value

string

Examples

# \donttest{
tbl <-
  df_adverse_events %>%
  tbl_ae(
    id = patient_id,
    ae = adverse_event,
    soc = system_organ_class,
    by = grade
  )
show_header_names(tbl)
#> 
#> ℹ As a usage guide, the code below re-creates the current column headers.
#> modify_header(
#>   label = '**Adverse Event**',
#>   stat_1 = '**1**',
#>   stat_2 = '**2**',
#>   stat_3 = '**3**',
#>   stat_4 = '**4**',
#>   stat_5 = '**5**'
#> )
#> 
#> 
#> Column Name   Column Header     
#> ------------  ------------------
#> label         **Adverse Event** 
#> stat_1        **1**             
#> stat_2        **2**             
#> stat_3        **3**             
#> stat_4        **4**             
#> stat_5        **5**             

inline_text(tbl, "Anaemia", column = stat_5)
#> [1] "3 (30)"
# }